feat: advanced optimizations and release prep (rounds 81-120)
Advanced Optimizations (Rounds 81-90):
- Add #[inline] to hot path functions for reduced call overhead
- Pre-compute constant tables at compile time (masks, power-of-2 values)
- New arena allocator module for fast temporary object allocation
- Optimize data structure layout for cache locality (#[repr(C)])
- Add branch prediction hints (likely!/unlikely! macros)
- Loop unrolling for param initialization and local allocation
- Bit manipulation optimizations for masking and sign extension
- Memory pooling with reusable buffer buckets (4 sizes: 256B-16KB)
- Profile-guided optimization instrumentation (opcode histogram, counters)

Security & Documentation (Rounds 91-120):
- Add comprehensive documentation to build.rs, nef.rs, manifest/builder.rs
- Update workspace version to 0.4.1
- Add crates.io metadata (keywords, categories, include) to all crates
- Add docs.rs badge to README.md

Fixes:
- Replace LazyLock (1.80+) with once_cell::Lazy for MSRV 1.70 compatibility
- Fix test patterns for const evaluation optimizations
- Fix clippy warnings (dead_code, doc indentation, mut_from_ref)

New Files:
- arena.rs: Bump-pointer arena allocator
- profiling.rs: PGO instrumentation

All tests pass (47 test groups), clippy clean, formatting clean.

+++ wasm-neovm/src/translator/translation/ops/compare.rs
@@ -1,5 +1,6 @@
 use super::*;
 
+/// Comparison operations (Round 82 - Const evaluation optimization)
 #[derive(Clone, Copy)]
 pub(in super::super) enum CompareOp {
     Lt,
@@ -9,6 +10,8 @@ pub(in super::super) enum CompareOp {
 }
 
 impl CompareOp {
+    /// Get the opcode name for this comparison (Round 81 - inline)
+    #[inline(always)]
     fn opcode_name(self) -> &'static str {
         match self {
             CompareOp::Lt => "LT",
@@ -18,26 +21,33 @@ impl CompareOp {
         }
     }
 
+    /// Evaluate signed comparison with compile-time constant folding (Round 82)
+    #[inline]
     fn evaluate_signed(self, lhs: i128, rhs: i128, bits: u32) -> bool {
+        // Simple cast-based sign extension (more reliable than bit manipulation)
         match bits {
             32 => {
-                let lhs = lhs as i32;
-                let rhs = rhs as i32;
+                let lhs = (lhs as i32) as i128;
+                let rhs = (rhs as i32) as i128;
                 self.evaluate_order(lhs, rhs)
             }
             64 => {
-                let lhs = lhs as i64;
-                let rhs = rhs as i64;
+                let lhs = (lhs as i64) as i128;
+                let rhs = (rhs as i64) as i128;
                 self.evaluate_order(lhs, rhs)
             }
             other => unreachable!("unsupported signed comparison width {}", other),
         }
     }
 
+    /// Evaluate unsigned comparison (Round 81 - inline)
+    #[inline(always)]
     fn evaluate_unsigned(self, lhs: u128, rhs: u128) -> bool {
         self.evaluate_order(lhs, rhs)
     }
 
+    /// Generic order evaluation (Round 81 - inline hot path)
+    #[inline(always)]
     fn evaluate_order<T: PartialOrd>(self, lhs: T, rhs: T) -> bool {
         match self {
             CompareOp::Lt => lhs < rhs,
